+++ tests/Config/DebugSettings.test.ts
@@ -1,7 +1,7 @@
 import { DebugSettings } from '../../src/Config/DebugSettings';
 import { resetSettings, updateSettings } from '../../src/Config/Settings';
 import { Query } from '../../src/Query/Query';
-import { createTasksFromMarkdown } from '../TestHelpers';
+import { createTasksFromMarkdown } from '../TestingTools/TestHelpers';
 
 describe('DebugSettings', () => {
     afterEach(() => {
