+++ tests/CustomMatchers/CustomMatchersForFilters.ts
@@ -1,7 +1,7 @@
 import { diff } from 'jest-diff';
 import type { Task } from '../../src/Task/Task';
 import type { FilterOrErrorMessage } from '../../src/Query/Filter/FilterOrErrorMessage';
-import { fromLine } from '../TestHelpers';
+import { fromLine } from '../TestingTools/TestHelpers';
 import { TaskBuilder } from '../TestingTools/TaskBuilder';
 import type { StatusConfiguration } from '../../src/Statuses/StatusConfiguration';
 import { Status } from '../../src/Statuses/Status';
