+++ tests/Query/Filter/FolderField.test.ts
@@ -1,5 +1,5 @@
 import { FolderField } from '../../../src/Query/Filter/FolderField';
-import { SampleTasks, fromLine } from '../../TestHelpers';
+import { SampleTasks, fromLine } from '../../TestingTools/TestHelpers';
 
 describe('folder', () => {
     it('should provide access to the folder', () => {
