+++ tests/Query/Filter/RootField.test.ts
@@ -1,5 +1,5 @@
 import { RootField } from '../../../src/Query/Filter/RootField';
-import { SampleTasks, fromLine } from '../../TestHelpers';
+import { SampleTasks, fromLine } from '../../TestingTools/TestHelpers';
 
 describe('root', () => {
     it('should provide access to root folder name with trailing slash', () => {
