+++ tests/Scripting/ScriptingReference/CustomSorting/CustomSortingExamples.test.ts
@@ -4,7 +4,7 @@
 
 import moment from 'moment';
 import type { Task } from '../../../../src/Task/Task';
-import { SampleTasks, fromLine, fromLines } from '../../../TestHelpers';
+import { SampleTasks, fromLine, fromLines } from '../../../TestingTools/TestHelpers';
 import type { CustomPropertyDocsTestData, QueryInstructionLineAndDescription } from '../VerifyFunctionFieldSamples';
 import {
     verifyFunctionFieldFilterSamplesForDocs,
