+++ tests/TaskSerializer/DocsSamplesForTaskFormats.test.ts
@@ -6,7 +6,7 @@ import moment from 'moment';
 
 import { TASK_FORMATS, resetSettings, updateSettings } from '../../src/Config/Settings';
 import { verifyMarkdown, verifyMarkdownForDocs } from '../TestingTools/VerifyMarkdown';
-import { SampleTasks } from '../TestHelpers';
+import { SampleTasks } from '../TestingTools/TestHelpers';
 
 window.moment = moment;
