+++ tests/TestingTools/TestHelpers.ts
@@ -1,10 +1,10 @@
-import { Priority, Task } from '../src/Task/Task';
-import { TaskLocation } from '../src/Task/TaskLocation';
-import { PriorityTools } from '../src/lib/PriorityTools';
-import { Status } from '../src/Statuses/Status';
-import { StatusType } from '../src/Statuses/StatusConfiguration';
-import { Recurrence } from '../src/Task/Recurrence';
-import { TaskBuilder } from './TestingTools/TaskBuilder';
+import { Priority, Task } from '../../src/Task/Task';
+import { TaskLocation } from '../../src/Task/TaskLocation';
+import { PriorityTools } from '../../src/lib/PriorityTools';
+import { Status } from '../../src/Statuses/Status';
+import { StatusType } from '../../src/Statuses/StatusConfiguration';
+import { Recurrence } from '../../src/Task/Recurrence';
+import { TaskBuilder } from './TaskBuilder';
 
 export function fromLine({
     line,
